Fixing colour tilemap glitches
Tilemaps on PAWSv1 were stored as an array of tilenumbers and the attributes (colours,reflection).

PAWSv2 combines these into a single BRAM { rotate/reflect, action(2bits),tielnumber(6bit) } so that the same tilenumber and attributes are being read at the same time.

Changes made to update tilemap and video_memmap.

Space Invaders which is being used to test this now has extra background objects as a test.