Various Updates
Add in a video clock to try and meet timings due to the updated colour mapping.

Properly define the 10%,20% etc mappings, plus improve the mapping for the RRGGGBB green components.

Add in a utility to output GIF files (with a PAWSv2 palette from GIMP) to the pixel mappings for PAWSv2.

Redo the bitmaps from asteroids and maze to use the new colour mappings.