DMA + SDCARD + SOFTWARE TWEAKS
Add in a register in the DMA for an unsigned char, can be used by memset, and the new mode 4.

Add in a mode 4 for the DMA engine, fixed source address to increasing destination address, used by the SDCARD readsector + MEMSET.

SDCARD in io_memmap. Make the buffer pointers for the sdcard increase when read/written, and add in a reset flag for the buffer pointers. Used by the DMA engine when copying from the sdcard buffer to main memory, and from main memory to the sdcard buffer.

BIOS, update the SDCARD readsector to use the DMA engine.

PAWSlibrary, update SDCARD readsector to use the DMA engine. Add in a memset using the DMA engine.

crt0.c pass 0 and NULL to main to signal no parameters, enables DOOM to start. Use DMA engine to clear the bss.

DOOM  in r_draw.c move a few tables to BRAM to speed up some of the loops.

Recompile BIOS and all software after updates.