USe DMA for setting colourblittiles and spritebitmaps
Reconfigure the blitter, character blitter, and colour blitter writers to reset to 0 when tile number is written, and to auto increment when a line/pixel is written.

Reconfigure the sprite bitmap writer to reset to 0 when the sprite number is set and to auto increment when a pixel is written.

Allows the use of DMA for setting colour blitter and sprite tiles, plus simplifies writing of blitter and character blitter bitmaps.

Make changes to BIOS/VBIOS and library, plus recompile affected software.

Allows