tilemap work
Give the tile map the ability to scroll by 0 - 15 pixels when scrolling/wrapping.

Provide a memory register in video_memmap for the adjustment amount. Correctly output tm_active and tm_lastaction as was reading 0.

Adjust library so that a scroll/wrap number of pixels is sent to the tilemap, plus BIOS and VBIOS.

Recompile BIOS, VBIOS and software that uses the tilemap, asteroids, demo, spaceinvaders, outrun.