audio work
Adjust audio waveforms so as not to use 0. Ensure audio output is 0 when no sound being output.

Provide a "sample" mode for audio, uses the DOOM pc speaker sample format, a series of notes for a square wave to play for a given duration. Tone is mapped from 1-127 for DOOM to 1-63 for PAWS.

Tested with DOOM, makes sounds, but probably not in sync with the video.

Adjust the I/O memmap to provide storage and addresses to write to the left and right samples, upload via DMA with library.

Sample controller is in I/O memmap above the audio controller, and sends series of tone requests to the audio controller.