L0 instruction cache
Add in a small (8 instructions) cache in the CPU. Saves around 4 seconds on the DOOM demo to the armour pickup. Use a clock for this cache.

Save BRAM by attempting to use the DPR16X4 blocks, used in the L0 cache. Move audio sine table to DPR16X4 by treating as bram, not brom. Split background copper storage into smaller sections so is transferred to DPR16X4. Split CPU registers into HARTR 0 and 1, within the actual control block for a register.