BITMAP CPU memory access
Change bitmap memory access to be 16 bit, allows the wipe effect in DOOM to be used correctly when using the spare framebuffer as the rendering area.

DOOMTEST renders to the spare framebuffer area and then uses the pixelbuffer to transfer to the display buffer whilst remapping the colours.

L1 caches have been reduced to a maximum of 4096 entries to allow the bitmaps to be changed to 8 bit, from 7 bits.

TODO: Change all other pixels from 7 to 8 bits then redo the palettes.