PAWSv2 moved to 256 colours
All display units (except blue/white little terminal window), now run at 256 colours.

Default mode is RGBM ( RRGGGBBM ) where M adds the lower bits to R and B to make RRRGGGBBB, but with only 256 colours.

RGB mode is extended to RRRGGGBB, and greyscale is extended to 256 colours.

Third mode is presently not decided upon.

Demos will need to be recompiled, once the colour defines are completed in the library files, they all work, just the colours are a bit out.

DOOM is working at 256 colours, using the spare frame buffer as screen[0] for rendering the screen before displaying. The time to armour pickup is finally under 32 seconds, just!

BRAM has been moved to use a write mask when writing bytes, allowing a cycle to be removed, allowing fast stores to be activated for the bram as per the I/O memory.