CPU reorganisation
Reorganise the main CPU loop, providing a path through for each type of instruction. Reorganise the mapping of addresses from the memory control units to the memory controllers. Move back to the multi-block L0 caches. Reorganise the memory controllers depending upon the size of the load/store.

Bitmap - allow the removal of transparent from the bitmap layer, allowing 256 colours, mainly for the DOOM demo. Adjust the pixel block to allow the writing of all 256 colours. Provide a reset flag in the BIOS. Plus an activation flag in the library. Map into memory in video_memmap.

Move DOOM back to RGBM mode - remove 256 from 2^24 colour mode, as wasn't mapping correctly.

Minor optimisation in nanojpeg to use pack and bset.