yosys-0.17 changes
Changes to the L0 change to build and meet timings with yosys-0.17. Shrink L0 blocks to 32 entries, but double number of blocks to achieve the same size of L0 cache. Meets timings, tested, no regressions found.

Change the size calculation in the blitter.

Start to move the outrun-demo the RGBM mode. tilemaps still need to be changed.

DOOMTEST remove the netgame code, as PAWS is not network capable, so excess code.