Sort out clocking for palette....
...to meet timings.

Use a 50MHz clock just for the palette lookups, double the pixel clock so that it can be done in the same pixel cycle.

Use a simple palette lookup algorithm in the multiplex display to move it out of inline code.

Simplify terminal scrolling/clearing.

Remove input! from the character rom and tile storage.

BUILD using a large L0 cache that still meets timing.