FPU rounding
Pass function3 to intlong2floatdouble and to floatdoubletointlong to all rounding (only RNE, RTZ) enabled. In FPU64 select either roudning flag from CSR or from instruction if dynamic and pass to the round and combine unit (only RNE, RTZ).

Fixes issue with D-STONE and allows floating point result to print correctly for cycles per instruction, and QUAKE to allow the size of the ehap to correctly display.

Start import of IMGUI from learnfpga.