FPU fixes and optimisations
CONTINUE WITH COMMENTING THE FPU CODE, still to do, the single cycle operations and the calculation controller.

typeF unit - add comments for inputs/outputs and within the unit.

intlong2floatdouble unit - add comments for inputs/outputs and within the unit. Add a notZERO check on the input. Add a roundingbit extraction and zero it if round to zero.

floatdouble2intlong unit - add comments for inputs/outputs and within the unit. Add a roundingbit extraction and zero it if round to zero.

changeprecision unit - add comments for inputs/outputs and within the unit.

donormal and fastnormal units - add comments for inputs/outputs and within the units.

doroundcombine overflow and overflow units - add comments for inputs/outputs and within the units. Add a rounding bit extraction (lsb) and zero it if round to zero. Add in signed zero and infinity results.

floataddsub dofloataddsub and equaliseexpaddsub units - add comments for inputs/outputs and within the units. Add sign extractions and pass to dofloataddsub unit.

floatmultiply units - add comments for inputs/outputs and within the unit.

floatdivide and dofloatdivide units - add comments for inputs/outputs and within the units.  Register the left and right aligned dividend and divisor and pass to the dofloatdivide unit.

dofloatsqrt and floatsqrt units - add comments for inputs/outputs and within the units. Reduce the number of bits processed to 53 for double, 24 for float ( 1 + length of fraction ). Reduce returned square root to 64 bits.