README.md updates + FPU64.si improvements
Add in some extra screenshots to README.md

FPU64.si

Split out from fpuSINGLECYCLE the units to do floatclass floateqltle floatminmax floatsign, and use within fpuSINGLECYCLE. Document inputs/outputs and within the units.

Simplify and document within the floatcalc unit. Register passing of result bitstream from addsub and division to the normalisation unit. Register the passing of the rounding mode flags from CSR or instruction to the rounding and combining unit.

Split out from floatdivide the preparation of the inputs to the floating point division. Rename sigBZERO to powTWO.

Split out from floatsquareroot the preparation of the input to floating point squareroot. Add in detection of power of 2 squareroots and return 1 or squareroot(2) if even or odd exponent. Correct result will be generated by halving the input exponent.