Move floating point changeprecision to repacked source register
cpu_functionblocks.si for changepreciision, use the opposite of function7[0,1] to classify the floating point register, and to expand the register into double bitfields (function7[0,1] for change precision indicates the destination type, so the register will be the other type).

CPU.si pass typeAopp and sourceReg1Fxopp (the register in double bitfields) to cpuexecuteFPU.

FPU64.si use the float in double bitfields for the change float <-> double operation.