RV64 floating point fused-multiply-add
Replace re-use of the floating-point multiplication and addition units to do fused-multiply-add operations with a dedicated floatfused unit.More accurate as the results of the multiplication are not rounded before use in the addition/subtraction operation.

Saves cycles as the rounding and combining step is eliminated after the multiplication

Chnage wait-state system in floataddsub floatmultiply.