zca zcb instruction expansion
NOT YET TESTED.

Implement zca, zcb, zcd as the compressed instruction support. The expansion of the additional compressed instructions has yet to be tested, due to lack of GCC support (I'm still on 12.2.0 due to lag in ArchLinux).

Simplify various other expansions.

NEXT STEP: remove the trailing 2 bits in the expansion, as not required in each of the 3 modules.