Refactoring floating-point unit and tilemaps
Move a lot of common items into cpu_functionblocks.si from FPU64.si

Generate the zerofraction flags for a and b , and pass to fclass, floatcompare, floatdivide and floatsquareroot. Used to determine true zeros, and power of 2. Generate common tests, a OR b, a AND b, INF, NAN, ZERO and pass to units as appropriate.

Refactor units to use the a OR b, a AND b INF, NAN, ZERO flags, and the zerofraction flags.

Use the do_float_multiply unit to generate the sign for products and quotients.

Make each calculation unit self contained, doing normalising, rounding and combining within the unit. Easier to follow the logic within each unit as the there is a natural cascade from one unit to the next within each calculation unit.

Tile maps are now 64x64 tiles, with a 2 tile border on the left and top side. Simplifies the logic, and makes them a little more flexible.