Cycle optimisations
DSTONE previous best was about 7.3 cycles per operation on average, now reduced to 6.320 cycles per operation on average, by moving the ALU out of CPUFAST in CPU.si

ALU.si add in support for the Zicond extension, not tested as archlinux does not yet support gcc14 for riscv. Implement state machine for integer division.

GPU is now clocked at 50MHz, changes made in video_memmap to remove the 25MHz clock bridging.

Changes in cpu_functionblocks.si to use floating-point registers direct from the register block for the repackaging. This allows the DELAY cycle in CPU.si for floating-point operations to only be used for calculations.

Remove a wait state for writing to I/O registers in the dma_bytes_to_io block.

FPU64.si implement state machines for division and squareroot.

Remove a wait state when reading from the bitmap or charactermap, now that the GPU is clocked at 50MHz.