Screen Layering Changes
As per sprites, the tilesets used for the tilemaps had a port0, used for drawing the screen, and a port1 used for writing. Double the number of tilemaps, 0,1 use tileset 0, 2,3 use tileset 1.

Merge the upper and lower sprites into a single sprite layer, but group into groups of 16, with 0-15 being in layer 0, 16-31 being in layer 1, 32-47 being in layer 2 and 48-63 being in layer 3.

Allow collision detection directly between all 64 sprites.

The sprite tileset system has been made more flexible, allowing for the standard 8x16x16 tileset for each pair of sprites, but also allowing for a 2x32x32 tileset for each pair of sprites.

Chnage the display multiplexer to allow for arbitary arrangement of the display layers, including both bitmaps displayed simultaneously, at differing layer positions (previously bitmaps could not be decoupled).

Has involved major changes to the display memory maps to allow for the registers for the 64 sprites to fit within 1 block.

There are changes required to the BIOS, software library to support these changes.

OUTRUN demo has been used as the main test system for the layering changes, adding in extra layers to the city scapes, and increasing the number of cars shown in the background. The Ferrari is now drawn using 32x32 tileset sprites.

Most of the software packages need updating, ASTROIDS has been modified, and appears to work using the 64 sprite system. Other packages still need a proper conversion.

Uploading now, as a further change to the audio system is planned, allowing 1 pcm and 3 tone generators LEFT and RIGHT, via time-multiplexing of the PWM.